deps/v8/test/inspector/debugger/wasm-gc-anyref-expected.txt

Test inspection of Wasm anyref objects

Running test: test
Calling instantiate function.
Waiting for wasm script to be parsed.
Got wasm script!
Module instantiated.
Tables populated.
Setting breakpoint
{
    columnNumber : 316
    lineNumber : 0
    scriptId : <scriptId>
}
Paused:
Script wasm://wasm/4ed17ea6 byte offset 316: Wasm opcode 0x01 (kExprNop)
Scope:
at $main (0:316):
 - scope (wasm-expression-stack):
   stack:
 - scope (local):
   $anyref_local: Struct ((ref $canon4))
   $anyref_local2: Array ((ref $canon5))
   $anyref_local_i31: 30 (anyref)
   $anyref_local_null: null (anyref)
 - scope (module):
   instance: exports: "exported_ref_table" (Table), "exported_func_table" (Table), "exported_i31_table" (Table), "fill_tables" (Function), "main" (Function)
   module: Module
   functions: "$my_func": (Function), "$fill_tables": (Function), "$main": (Function)
   globals: "$global0": function $my_func() { [native code] } (funcref)
   tables:
      $import.any_table: 0: Array(2) (anyref), 1: 321 (anyref), 2: null (anyref), 3: null (anyref)
      $import.func_table: 0: function () { [native code] } (funcref), 1: function $my_func() { [native code] } (funcref), 2: null (funcref)
      $exported_ref_table: 0: Struct ((ref $canon4)), 1: Array ((ref $canon5)), 2: 30 (anyref), 3: null (anyref)
      $exported_func_table: 0: function external_fct() { [native code] } (funcref), 1: function $my_func() { [native code] } (funcref), 2: null (funcref)
      $exported_i31_table: 0: 123456 (i31ref), 1: -123 (i31ref), 2: null (i31ref)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

exports.main returned. Test finished.
